abstract out createControl
