+++ src/modelInput/modelInput.js
@@ -1,3 +1,5 @@
+import { createControl } from './createControl';
+
 import { requestModelBuilders } from '../h2oProxy/requestModelBuilders';
 import { postModelInputValidationRequest } from '../h2oProxy/postModelInputValidationRequest';
 
@@ -8,37 +10,6 @@ export function modelInput() {
   const lodash = window._;
   const Flow = window.Flow;
   const H2O = window.H2O;
-  const createControl = (kind, parameter) => {
-    const _hasError = Flow.Dataflow.signal(false);
-    const _hasWarning = Flow.Dataflow.signal(false);
-    const _hasInfo = Flow.Dataflow.signal(false);
-    const _message = Flow.Dataflow.signal('');
-    const _hasMessage = Flow.Dataflow.lift(_message, message => {
-      if (message) {
-        return true;
-      }
-      return false;
-    });
-    const _isVisible = Flow.Dataflow.signal(true);
-    const _isGrided = Flow.Dataflow.signal(false);
-    const _isNotGrided = Flow.Dataflow.lift(_isGrided, value => !value);
-    return {
-      kind,
-      name: parameter.name,
-      label: parameter.label,
-      description: parameter.help,
-      isRequired: parameter.required,
-      hasError: _hasError,
-      hasWarning: _hasWarning,
-      hasInfo: _hasInfo,
-      message: _message,
-      hasMessage: _hasMessage,
-      isVisible: _isVisible,
-      isGridable: parameter.gridable,
-      isGrided: _isGrided,
-      isNotGrided: _isNotGrided,
-    };
-  };
   const createTextboxControl = (parameter, type) => {
     let isArrayValued;
     let isInt;
